[Trivial] Rename & regroup linearizability tests
Python tests and utility modules related to linearizability needed some renaming, in order to state their purpose more clearly:

test_skvbc_linearizability -> test_skvbc_history_tracker
test_skvbc_chaos -> test_skvbc_linearizability
util/skvbc_linearizabillity -> util/skvbc_history_tracker

Note: due to this "circular" renaming, in the PR test_skvbc_history_tracker.py appears as a new file, but it's simply renamed without code changes
